feat: Add emissionDate field to Payable creation
This commit adds the `emissionDate` field to the `create` method of the `Payable` entity in the `payable.ts` file. The `emissionDate` is now an optional property in the `PayableProps` type, allowing it to be passed when creating a new `Payable` instance. If no `emissionDate` is provided, the current date is used as the default value.